split tests. add tests.
