+++ .github/workflows/main.yml
@@ -25,6 +25,8 @@ jobs:
         python -m pip install --upgrade pip
         pip install -r requirements.txt
 
+        PYTHONPATH=. pytest tests
+
     - name: Test canary
       env:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
