+++ entrypoint.sh
@@ -15,19 +15,23 @@ pyenv virtualenv $INPUT_PYTHON_VERSION venv
 pyenv activate venv
 
 echo "🔥🔥🔥🔥🔥Running security check🔥🔥🔥🔥🔥🔥"
-pip install bandit
+pip -r requirements.txt
 mkdir -p $GITHUB_WORKSPACE/output
 touch $GITHUB_WORKSPACE/output/security_report.txt
 bandit -r $INPUT_PROJECT_PATH -o $GITHUB_WORKSPACE/output/security_report.txt -f 'txt'
+BANDIT_STATUS="$?"
 
-if [ $? -eq 0 ]; then
+python annotations.py -r $INPUT_PROJECT_PATH
+
+if [ $BANDIT_STATUS -eq 0 ]; then
     echo "🔥🔥🔥🔥Security check passed🔥🔥🔥🔥"
-else
-    echo "🔥🔥🔥🔥Security check failed🔥🔥🔥🔥"
-    cat $GITHUB_WORKSPACE/output/security_report.txt
-    if $INPUT_IGNORE_FAILURE; then
-        exit 0
-    else
-        exit 1
-    fi
+    exit 0
 fi
+
+echo "🔥🔥🔥🔥Security check failed🔥🔥🔥🔥"
+cat $GITHUB_WORKSPACE/output/security_report.txt
+if $INPUT_IGNORE_FAILURE; then
+    exit 0
+fi
+
+exit $BANDIT_STATUS
