--- main.py
@@ -1,3 +1,0 @@
-import Popen
-
-self.process = subprocess.Popen('/bin/echo', shell=True)
